refactor: Rename algorithm classes and folders
* Rename NVSemiParametricGEstimationPostWidder → NVEstimationDensityPW
* Rename SemiParametricGEstimationPostWidder → NMVEstimationDensityPW
* Rename folders in src/procedures
* Rename tests/algorithms → tests/procedures

lgorithms/semiparametric_sigma_estimation/__init__.py -> tests/procedures/nm_procedures/semiparametric_sigma_estimation/__init__.py
