+++ src/estimators/abstract_estimator.py
@@ -14,7 +14,7 @@ class AbstractEstimator:
         algorithm_name: A string indicating chosen algorithm.
         params: A dictionary of algorithm parameters.
         estimate_result: Estimation result.
-        _registry: Registry that contains classes of all algorithms.
+        _registry: Registry that contains classes of all procedures.
         _purpose: Defines purpose of algorithm, one of the registry key.
     """
 
@@ -47,7 +47,7 @@ class AbstractEstimator:
         return self
 
     def get_available_algorithms(self) -> list[str]:
-        """Get all algorithms that can be used for current estimator class"""
+        """Get all procedures that can be used for current estimator class"""
         return [key[0] for key in self._registry.register_of_names.keys() if key[1] == self._purpose]
 
     def estimate(self, sample: _typing.ArrayLike) -> EstimateResult:
