+++ src/procedures/semiparametric/nm_semiparametric/sigma_estimation_eigenvalue_based.py
