+++ src/procedures/semiparametric/nvm_semiparametric/g_estimation_post_widder.py
@@ -14,7 +14,7 @@ N_DEFAULT_VALUE = 2
 X_DATA_DEFAULT_VALUE = [1.0]
 
 
-class SemiParametricGEstimationPostWidder:
+class NMVEstimationDensityPW:
     """Estimation of mixing density function g (xi density function) of NVM mixture represented in classical form Y =
     alpha + mu*xi + sigma*sqrt(xi)*N, where alpha = 0 and mu, sigma are given.
