+++ src/procedures/semiparametric/statistical_procedure.py
@@ -4,7 +4,7 @@ import numpy as np
 from src.estimators.estimate_result import EstimateResult
 
 class StatisticalProcedure(Protocol):
-    """Protocol for statistical algorithms that compute estimates from data.
+    """Protocol for statistical procedures that compute estimates from data.
 
     """
     def compute(self, sample: np.ndarray) -> EstimateResult:
